build: generate public API wrappers from source
