+++ .github/workflows/ci.yml
@@ -88,6 +88,9 @@ jobs:
       - name: Build
         run: nr build
 
+      - name: Verify generated API is committed
+        run: nr check:api
+
       - name: Typecheck
         run: nr typecheck
