+++ node/package.json
@@ -20,6 +20,7 @@
     "test:ferriki-compat:adapters": "pnpm run build:compat && pnpm -C ferriki build:native && FERRIKI_BACKEND=rust vitest compat/upstream/shiki/packages/transformers/test compat/upstream/shiki/packages/twoslash/test --exclude compat/upstream/shiki/packages/twoslash/test/markdown-it.test.ts --run --maxWorkers 1 --no-file-parallelism",
     "test:ferriki-compat:colorized-brackets": "pnpm run build:compat && pnpm -C ferriki build:native && node ./compat/harness/run-colorized-brackets-compat.mjs",
     "check:errors": "node ./scripts/check-ferriki-errors.mjs",
+    "check:api": "node ./scripts/check-generated-api.mjs",
     "check:boundary": "node ./scripts/check-native-boundary.mjs",
     "check:release": "node ./scripts/check-release-workflow.mjs",
     "check:transformers": "node ./scripts/check-ferriki-transformers.mjs",
